+++ lambdarpc-core/src/main/kotlin/io/lambdarpc/transport/grpc/SimpleGrpcConnection.kt
@@ -1,10 +1,8 @@
-package io.lambdarpc.transport.grpc.service
+package io.lambdarpc.transport.grpc
 
 import io.grpc.ManagedChannel
 import io.grpc.ManagedChannelBuilder
 import io.lambdarpc.transport.Connection
-import io.lambdarpc.transport.grpc.InMessage
-import io.lambdarpc.transport.grpc.OutMessage
 import io.lambdarpc.utils.Endpoint
 import kotlinx.coroutines.flow.Flow
