+++ lambdarpc-core/src/main/kotlin/io/lambdarpc/transport/serialization/Utils.kt
@@ -1,4 +1,4 @@
-package io.lambdarpc.transport.grpc.serialization
+package io.lambdarpc.transport.serialization
 
 import com.google.protobuf.ByteString
 import io.lambdarpc.ExecutionException
