--- lambdarpc-core/src/test/kotlin/io/lambdarpc/ExceptionsTest.kt
@@ -1,4 +1,0 @@
-package io.lambdarpc
-
-// TODO
-class ExceptionsTest
