--- lambdarpc-core/src/test/kotlin/io/lambdarpc/MultipleServices.kt
@@ -1,4 +1,0 @@
-package io.lambdarpc
-
-// TODO
-class MultipleServices
